+++ ipaserver/dnssec/bindmgr.py
@@ -14,7 +14,7 @@ from ipapython.dn import DN
 from ipapython import ipa_log_manager, ipautil
 from ipaplatform.paths import paths
 
-from ipapython.dnssec.temp import TemporaryDirectory
+from ipaserver.dnssec.temp import TemporaryDirectory
 
 time_bindfmt = '%Y%m%d%H%M%S'
