+++ ipaserver/dnssec/ldapkeydb.py
@@ -13,12 +13,12 @@ from ipapython.dn import DN
 from ipapython import ipaldap
 from ipapython import ipa_log_manager
 
-from ipapython.dnssec.abshsm import (
+from ipaserver.dnssec.abshsm import (
     attrs_name2id,
     AbstractHSM,
     bool_attr_names,
     populate_pkcs11_metadata)
-from ipapython import p11helper as _ipap11helper
+from ipaserver import p11helper as _ipap11helper
 import uuid
 
 def uri_escape(val):
